day 2
entirely unexciting second part solution: my initial implementation was
iterating with first and rest and simply iterating on to the next
element keeping the previous-previous element to compare it to. This
failed when the first element was the one that needed dropping. Fixing
that still yielded 4 cases that were incorrect but I would rather do
Christmas shopping than deal with those.